Refactor and separate code parts
