+++ lantern_pq/src/setup.rs
@@ -1,0 +1,70 @@
+use lantern_logger::Logger;
+use lantern_utils::quote_ident;
+use postgres::Transaction;
+
+use crate::AnyhowVoidResult;
+
+// Will create a codebook table add neccessary indexes and add PQVEC column into target table
+pub fn setup_tables<'a>(
+    transaction: &mut Transaction<'a>,
+    full_table_name: &str,
+    codebook_table_name: &str,
+    pq_column_name: &str,
+    logger: &Logger,
+) -> AnyhowVoidResult {
+    transaction.batch_execute(&format!(
+        "
+             CREATE TABLE {codebook_table_name} (subvector_id INT, centroid_id INT, c REAL[]);
+             ALTER TABLE {full_table_name} ADD COLUMN {pq_column_name} PQVEC;
+             CREATE INDEX ON {codebook_table_name} USING BTREE(subvector_id, centroid_id);
+             CREATE INDEX ON {codebook_table_name} USING BTREE(centroid_id);
+        ",
+        codebook_table_name = quote_ident(&codebook_table_name),
+        pq_column_name = quote_ident(&pq_column_name)
+    ))?;
+    logger.info(&format!(
+        "{codebook_table_name} table and {pq_column_name} column created successfully"
+    ));
+    Ok(())
+}
+
+// Setup triggers to autoamtically compress new inserted/updated vectors
+pub fn setup_triggers<'a>(
+    transaction: &mut Transaction<'a>,
+    full_table_name: &str,
+    codebook_table_name: &str,
+    pq_column: &str,
+    column: &str,
+    distance_metric: &str,
+    splits: usize,
+) -> AnyhowVoidResult {
+    // Setup triggers for new data
+    let name_hash = md5::compute(format!("{}{}", full_table_name, pq_column));
+    let insert_trigger_name = format!("_pq_trigger_in_{:x}", name_hash);
+    let update_trigger_name = format!("_pq_trigger_up_{:x}", name_hash);
+    let trigger_fn_name = format!("_set_pq_col_{:x}", name_hash);
+
+    transaction.batch_execute(&format!("
+      DROP TRIGGER IF EXISTS {insert_trigger_name} ON {full_table_name};
+      DROP TRIGGER IF EXISTS {update_trigger_name} ON {full_table_name};
+
+      CREATE OR REPLACE FUNCTION {trigger_fn_name}()
+          RETURNS trigger
+          LANGUAGE plpgsql AS
+      $body$
+        BEGIN
+          IF NEW.{column} IS NULL THEN
+            NEW.{pq_column} := NULL;
+          ELSE
+            NEW.{pq_column} := _lantern_internal.compress_vector(NEW.{column}, {splits}, {codebook_table_name}::regclass, '{distance_metric}');
+          END IF;
+          RETURN NEW;
+        END
+      $body$;
+
+      CREATE TRIGGER {insert_trigger_name} BEFORE INSERT ON {full_table_name} FOR EACH ROW EXECUTE FUNCTION {trigger_fn_name}();
+      CREATE TRIGGER {update_trigger_name} BEFORE UPDATE OF {column} ON {full_table_name} FOR EACH ROW EXECUTE FUNCTION {trigger_fn_name}();
+
+    ", pq_column=quote_ident(pq_column), column=quote_ident(column), codebook_table_name=quote_ident(codebook_table_name)))?;
+    Ok(())
+}
